backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
SQLAlchemy==2.0.38
alembic==1.15.1
psycopg[binary]==3.2.6
pydantic-settings==2.8.1
python-dotenv==1.0.1
grpcio==1.62.3
grpcio-tools==1.62.3
protobuf==4.25.8
